test: consolidate ModeWidget tests and move statistics tests to __tests__
