+++ js/widgets/__tests__/modewidget.test.js
@@ -1,72 +1,186 @@
 /**
  * MusicBlocks v3.6.2
  *
- * @author Alok Dangre
+ * @author Ashutosh Kumar
  *
- * @copyright 2026 Alok Dangre
+ * @copyright 2026 Ashutosh Kumar
  *
  * @license
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-describe("ModeWidget._calculateMode logic", () => {
-    /**
-     * Extracted pure logic from ModeWidget._calculateMode
-     * Converts a 12-element boolean array (representing selected semitones)
-     * into an array of intervals (steps between selected notes).
-     */
-    const calculateMode = selectedNotes => {
-        const currentMode = [];
-        let j = 1;
-        for (let i = 1; i < 12; i++) {
-            if (selectedNotes[i]) {
-                currentMode.push(j);
-                j = 1;
-            } else {
-                j += 1;
-            }
-        }
-        currentMode.push(j);
-        return currentMode;
-    };
+const ModeWidget = require("../modewidget.js");
 
-    it("should return major scale intervals [2,2,1,2,2,2,1] for C major pattern", () => {
-        // C major: C(0), D(2), E(4), F(5), G(7), A(9), B(11)
-        const selectedNotes = [
-            true, // 0: C (root)
-            false, // 1: C#
-            true, // 2: D
-            false, // 3: D#
-            true, // 4: E
-            true, // 5: F
-            false, // 6: F#
-            true, // 7: G
-            false, // 8: G#
-            true, // 9: A
-            false, // 10: A#
-            true // 11: B
+// --- 1. Global Mocks ---
+global._ = msg => msg;
+global.platformColor = {
+    selectorBackground: "#FFFFFF",
+    selectorBackgroundHOVER: "#EEEEEE",
+    modeWheelcolors: ["#FF0000", "#00FF00"],
+    orange: "#FFA500"
+};
+global.DEFAULTVOICE = "piano";
+global.last = arr => arr[arr.length - 1];
+
+// Mock utils
+global.docById = jest.fn().mockImplementation(id => ({
+    style: {},
+    innerHTML: "",
+    appendChild: jest.fn(),
+    rows: [{ cells: [{ innerHTML: "" }] }, { cells: [{ innerHTML: "" }] }],
+    insertRow: jest.fn().mockReturnValue({
+        insertCell: jest.fn().mockReturnValue({
+            style: {},
+            innerHTML: ""
+        })
+    })
+}));
+
+global.getNote = jest.fn().mockReturnValue(["C", "4"]);
+global.keySignatureToMode = jest.fn().mockReturnValue(["C", "ionian"]);
+global.normalizeNoteAccidentals = jest.fn().mockImplementation(n => n);
+global.MUSICALMODES = {
+    ionian: [2, 2, 1, 2, 2, 2, 1]
+};
+
+// Mock slicePath
+global.slicePath = jest.fn().mockReturnValue({
+    DonutSlice: jest.fn(),
+    DonutSliceCustomization: jest.fn().mockReturnValue({})
+});
+
+// Mock wheelnav
+global.wheelnav = jest.fn().mockImplementation(() => ({
+    raphael: {},
+    colors: [],
+    slicePathFunction: null,
+    slicePathCustom: {},
+    sliceSelectedPathCustom: {},
+    sliceInitPathCustom: {},
+    navItems: [],
+    createWheel: jest.fn().mockImplementation(function (labels) {
+        this.navItems = labels.map(() => ({
+            navItem: {
+                hide: jest.fn(),
+                show: jest.fn()
+            },
+            navigateFunction: null
+        }));
+    })
+}));
+
+// Mock Window
+if (typeof window === "undefined") {
+    global.window = {};
+}
+window.innerWidth = 1024;
+window.widgetWindows = {
+    windowFor: jest.fn().mockReturnValue({
+        clear: jest.fn(),
+        show: jest.fn(),
+        destroy: jest.fn(),
+        updateTitle: jest.fn(),
+        addButton: jest.fn().mockImplementation(() => {
+            const btn = {};
+            btn.onclick = jest.fn();
+            return btn;
+        }),
+        getWidgetBody: jest.fn().mockReturnValue({
+            append: jest.fn(),
+            getElementsByTagName: jest.fn().mockReturnValue([
+                {
+                    style: {},
+                    setAttribute: jest.fn()
+                }
+            ])
+        }),
+        getWidgetFrame: jest.fn().mockReturnValue({ offsetHeight: 500 }),
+        getDragElement: jest.fn().mockReturnValue({ offsetHeight: 20 }),
+        isMaximized: jest.fn().mockReturnValue(false),
+        sendToCenter: jest.fn()
+    })
+};
+
+// Mock Document
+if (typeof document === "undefined") {
+    global.document = {};
+}
+document.createElement = jest.fn().mockImplementation(tag => ({
+    style: {},
+    innerHTML: "",
+    append: jest.fn(),
+    insertRow: jest.fn().mockReturnValue({
+        insertCell: jest.fn().mockReturnValue({
+            style: {},
+            innerHTML: ""
+        })
+    }),
+    getElementById: jest.fn().mockReturnValue({ src: "" })
+}));
+document.getElementById = document.createElement; // For internal usage
+
+describe("ModeWidget", () => {
+    let modeWidget;
+    let mockActivity;
+
+    beforeEach(() => {
+        mockActivity = {
+            logo: {
+                modeBlock: 1,
+                resetSynth: jest.fn(),
+                synth: {
+                    trigger: jest.fn(),
+                    stop: jest.fn()
+                }
+            },
+            turtles: {
+                ithTurtle: jest.fn().mockReturnValue({
+                    singer: { keySignature: ["C"] }
+                })
+            },
+            blocks: {
+                blockList: []
+            },
+            storage: {},
+            textMsg: jest.fn(),
+            hideMsgs: jest.fn(),
+            errorMsg: jest.fn(),
+            refreshCanvas: jest.fn()
+        };
+
+        modeWidget = new ModeWidget(mockActivity);
+    });
+
+    test("should calculate major scale intervals correctly", () => {
+        modeWidget._selectedNotes = [
+            true,
+            false,
+            true,
+            false,
+            true,
+            true,
+            false,
+            true,
+            false,
+            true,
+            false,
+            true
         ];
-        expect(calculateMode(selectedNotes)).toEqual([2, 2, 1, 2, 2, 2, 1]);
+
+        expect(modeWidget._calculateMode()).toEqual([2, 2, 1, 2, 2, 2, 1]);
     });
 
-    it("should return chromatic intervals when all notes selected", () => {
-        const selectedNotes = Array(12).fill(true);
-        expect(calculateMode(selectedNotes)).toEqual([1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1]);
+    test("should return all 1s for chromatic scale", () => {
+        modeWidget._selectedNotes = Array(12).fill(true);
+
+        expect(modeWidget._calculateMode()).toEqual(Array(12).fill(1));
     });
 
-    it("should return [12] when only root note is selected", () => {
-        const selectedNotes = [
+    test("should return [12] when only root is selected", () => {
+        modeWidget._selectedNotes = [
             true,
             false,
             false,
@@ -80,6 +194,168 @@ describe("ModeWidget._calculateMode logic", () => {
             false,
             false
         ];
-        expect(calculateMode(selectedNotes)).toEqual([12]);
+
+        expect(modeWidget._calculateMode()).toEqual([12]);
+    });
+
+    test("should handle empty selectedNotes safely", () => {
+        modeWidget._selectedNotes = [];
+
+        const result = modeWidget._calculateMode();
+
+        expect(result).toBeDefined();
+    });
+
+    test("should handle malformed selectedNotes safely", () => {
+        modeWidget._selectedNotes = [true, undefined, false, null, true];
+
+        const result = modeWidget._calculateMode();
+
+        expect(Array.isArray(result)).toBe(true);
+    });
+
+    test("undo should not crash when stack is empty", () => {
+        modeWidget._undoStack = [];
+        modeWidget._undo();
+
+        expect(modeWidget._selectedNotes).toBeDefined();
+    });
+
+    test("should return correct playing status", () => {
+        modeWidget._playing = false;
+        expect(modeWidget._playingStatus()).toBe(false);
+
+        modeWidget._playing = true;
+        expect(modeWidget._playingStatus()).toBe(true);
+    });
+
+    test("should save and undo state correctly", () => {
+        modeWidget._selectedNotes = Array(12).fill(false);
+        modeWidget._selectedNotes[0] = true;
+
+        modeWidget._saveState();
+        modeWidget._selectedNotes[1] = true;
+
+        modeWidget._undo();
+
+        expect(modeWidget._selectedNotes[1]).toBe(false);
+    });
+
+    test("should clear all notes except root", () => {
+        modeWidget._selectedNotes = Array(12).fill(true);
+
+        modeWidget._clear();
+
+        expect(modeWidget._selectedNotes[0]).toBe(true);
+        for (let i = 1; i < 12; i++) {
+            expect(modeWidget._selectedNotes[i]).toBe(false);
+        }
+    });
+
+    test("should trigger synth when playing a note", () => {
+        modeWidget._playNote(0);
+
+        expect(mockActivity.logo.synth.trigger).toHaveBeenCalled();
+    });
+
+    test("should initialize correctly", () => {
+        expect(global.wheelnav).toHaveBeenCalledTimes(3); // noteWheel, playWheel, etc.
+        expect(global.keySignatureToMode).toHaveBeenCalled();
+        expect(mockActivity.textMsg).toHaveBeenCalled();
+    });
+
+    test("should handle Play/Stop button toggle", () => {
+        jest.useFakeTimers();
+        const widgetWindow = window.widgetWindows.windowFor();
+        const playBtnMock = widgetWindow.addButton.mock.results[0].value;
+
+        // Start
+        expect(modeWidget._playing).toBe(false);
+        playBtnMock.onclick();
+        expect(modeWidget._playing).toBe(true);
+        expect(mockActivity.logo.resetSynth).toHaveBeenCalled();
+
+        // Stop
+        playBtnMock.onclick();
+        expect(modeWidget._playing).toBe(false);
+        jest.useRealTimers();
+    });
+
+    test("should rotate mode pattern right", () => {
+        // Setup: Need true at index 11 so that after 1 right shift (new[0] = old[11]), index 0 is true.
+        // This ensures the recursion stops after 1 rotation.
+        modeWidget._selectedNotes = [
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            true
+        ];
+
+        // Wait for rotation animations
+        jest.useFakeTimers();
+        modeWidget._rotateRight();
+
+        // Fast forward through animations (12 * ROTATESPEED)
+        jest.advanceTimersByTime(12 * 150 + 100);
+
+        expect(modeWidget._locked).toBe(false);
+        jest.useRealTimers();
+    });
+
+    test("should rotate mode pattern left", () => {
+        // Setup: Need true at index 1 so that after 1 left shift (new[0] = old[1]), index 0 is true.
+        modeWidget._selectedNotes = [
+            false,
+            true,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false
+        ];
+
+        jest.useFakeTimers();
+        modeWidget._rotateLeft();
+        jest.advanceTimersByTime(12 * 150 + 100);
+
+        expect(modeWidget._locked).toBe(false);
+        jest.useRealTimers();
+    });
+
+    test("should invert mode pattern", () => {
+        modeWidget._selectedNotes = [
+            true,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            false,
+            true,
+            false
+        ]; // C and B
+
+        jest.useFakeTimers();
+        modeWidget._invert(); // Inverts pairs around current axis
+        jest.advanceTimersByTime(6 * 150);
+
+        expect(modeWidget._locked).toBe(false);
+        jest.useRealTimers();
     });
 });
